feat: refactor module structure and rename transformers for clarity
