+++ src/quickapp/agent/_attachment_filter.py
