+++ src/quickapp/agent/chunk_processor.py
