+++ src/quickapp/agent/orchestrator.py
@@ -7,8 +7,8 @@ from pydantic.v1 import StrictStr
 
 from quickapp.agent.assistant_invoker import AssistantInvoker
 from quickapp.agent.models import TOOL_EXECUTION_HISTORY
-from quickapp.agent.processors.chunk_processor import ChunkProcessor
-from quickapp.agent.processors.tool_executor import ToolExecutor
+from quickapp.agent.chunk_processor import ChunkProcessor
+from quickapp.agent.tool_executor import ToolExecutor
 from quickapp.common import DeploymentUsage
 from quickapp.common.exceptions import OrchestratorExceedMaxIterationsException
 from quickapp.common.messages_mixin import MessagesMixin
